refact: remove redundant cdl directory
