+++ src/ast/actions_and_functions/action_declaration_term.rs
@@ -1,23 +1,24 @@
-use super::super::super::KeywordTerm;
-use super::super::super::PunctuationSignTerm;
-use super::super::IdentifierTerm;
-use super::super::ParametersBlockTerm;
-use super::super::ReturnsDeclarationTerm;
 use crate::ast::Location;
+
+use super::super::IdentifierTerm;
+use super::super::PunctuationSignTerm;
+use super::super::ReturnsDeclarationTerm;
+use super::ParametersBlockTerm;
+use crate::ast::basic::KeywordTerm;
 use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
-pub struct FunctionDeclarationTerm {
+pub struct ActionDeclarationTerm {
     #[prop]
     location: Location,
     #[subnode_prop]
-    function: Box<KeywordTerm>,
+    action: Box<KeywordTerm>,
     #[subnode_prop]
     identifier: Box<IdentifierTerm>,
     #[subnode_prop]
     parameters: Box<ParametersBlockTerm>,
     #[subnode_prop]
-    returns: Box<ReturnsDeclarationTerm>,
+    returns: Option<Box<ReturnsDeclarationTerm>>,
     #[subnode_prop]
     semicolumn: Box<PunctuationSignTerm>,
 }
