+++ src/ast/actions_and_functions/param_term.rs
@@ -1,13 +1,18 @@
-use super::super::super::KeywordTerm;
+use crate::ast::Location;
+
+use super::super::IdentifierTerm;
+use super::super::PunctuationSignTerm;
 use super::super::TypeReferenceTerm;
 use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
-pub struct ReturnsDeclarationTerm {
+pub struct ParamTerm {
     #[prop]
     location: Location,
     #[subnode_prop]
-    returns: Box<KeywordTerm>,
+    name: Box<IdentifierTerm>,
+    #[subnode_prop]
+    colon: Box<PunctuationSignTerm>,
     #[subnode_prop]
     type_reference: Box<TypeReferenceTerm>,
 }
