+++ src/ast/actions_and_functions/returns_declaration_term.rs
@@ -1,18 +1,13 @@
-use crate::ast::Location;
-
-use super::super::super::PunctuationSignTerm;
-use super::super::IdentifierTerm;
+use super::super::KeywordTerm;
 use super::super::TypeReferenceTerm;
 use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
-pub struct ParamTerm {
+pub struct ReturnsDeclarationTerm {
     #[prop]
     location: Location,
     #[subnode_prop]
-    name: Box<IdentifierTerm>,
-    #[subnode_prop]
-    colon: Box<PunctuationSignTerm>,
+    returns: Box<KeywordTerm>,
     #[subnode_prop]
     type_reference: Box<TypeReferenceTerm>,
 }
