--- src/ast/cdl/mod.rs
@@ -1,15 +1,0 @@
-mod actions_and_functions;
-mod entity;
-mod identifiers;
-mod modules;
-mod services;
-mod structure;
-mod types;
-
-pub use actions_and_functions::*;
-pub use entity::*;
-pub use identifiers::*;
-pub use modules::*;
-pub use services::*;
-pub use structure::*;
-pub use types::*;
